internal/features: move packages to internal
Fixes #37
