+++ internal/features/fastreflection/copied/file_info.go
