+++ internal/features/fastreflection/copied/naming.go
