+++ internal/features/fastreflection/proto_message.go
@@ -2,7 +2,7 @@ package fastreflection
 
 import (
 	"fmt"
-	"github.com/cosmos/cosmos-proto/features/fastreflection/copied"
+	"github.com/cosmos/cosmos-proto/internal/features/fastreflection/copied"
 	"github.com/cosmos/cosmos-proto/internal/generator"
 	"google.golang.org/protobuf/compiler/protogen"
 )
