+++ internal/features/protoc/genid/duration_gen.go
