+++ internal/features/protoc/genid/struct_gen.go
